Fix company redeem
Remove the specification of the company behaviour in the show
attendee, avoiding a GET from changing the attendee state.

Remove the plug Authorize in redeem_controller because there is 3
conditions in the create method where it is checked if the connection
belongs to a company or a manager, otherwise the connection is not
authorized and the redeem is not created. In this case the plug
is not needed.

Resolves: #93
